Converted the bot into a rack app
