+++ .env
@@ -1,0 +1,1 @@
+SLACK_API_TOKEN=xoxb-703864578774-701247943876-CjL5WUkXYcMF9SDfOtL31hcN
