+++ Gemfile
@@ -5,6 +5,11 @@ gem 'async-websocket', '~>0.8.0'
 
 gem 'sqlite3'
 
+# gem 'puma'
+# gem 'sinatra'
+# gem 'faraday_middleware', '0.10.1'
+
+
 group :development do
   gem 'irb'
   gem 'pry-byebug'
