+++ slack-nfl-bot.rb
@@ -1,0 +1,4 @@
+require 'slack-ruby-bot'
+require 'slack-nfl-bot/version'
+require 'slack-nfl-bot/commands'
+require 'slack-nfl-bot/app'
