+++ slack-nfl-bot/about.rb
@@ -1,0 +1,6 @@
+module SlackNFLBot
+  ABOUT = <<-ABOUT
+    #{SlackNFLBot::VERSION}
+    https://github.com/pcg79/nflbot
+  ABOUT
+end
