+++ slack-nfl-bot/app.rb
@@ -1,0 +1,4 @@
+module SlackNFLBot
+  class App < SlackRubyBot::App
+  end
+end
