+++ slack-nfl-bot/commands.rb
@@ -1,0 +1,3 @@
+require 'slack-nfl-bot/commands/nfl_bot'
+require 'slack-nfl-bot/commands/ping'
+require 'slack-nfl-bot/commands/about'
