+++ slack-nfl-bot/version.rb
@@ -1,0 +1,3 @@
+module SlackNFLBot
+  VERSION = '0.0.1'
+end
